Fix integer overflow when creating large times

+++ jtimeseries/src/main/java/com/od/jtimeseries/util/time/AbstractTimePeriod.java
@@ -26,9 +26,9 @@ package com.od.jtimeseries.util.time;
  */
 class AbstractTimePeriod implements TimePeriod {
 
-    private int lengthInMillis;
+    private long lengthInMillis;
 
-    public AbstractTimePeriod(int lengthInMillis) {
+    public AbstractTimePeriod(long lengthInMillis) {
         this.lengthInMillis = lengthInMillis;
     }
