+++ jtimeseries/src/main/java/com/od/jtimeseries/util/time/Minute.java
@@ -28,7 +28,7 @@ class Minute extends AbstractTimePeriod {
 
     private static final int millis = 60 * 1000;
 
-    public Minute(int mins) {
+    public Minute(long mins) {
         super(mins * millis);
     }
